+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Last Changes
 
+- [#16](https://github.com/LaxarJS/ax-details-layer-widget/issues/16): by default, scroll to top when opened
+    + **BREAKING CHANGE:** see ticket for details
+
 
 ## v1.0.0-alpha.2